requirements.txt
PyQt5
openpyxl
Pillow
pynput
pywin32
pyinstaller